gestion version

+++ ObjectDB.js
@@ -1,19 +1,20 @@
 (function() {
     "use strict";
-  
-    var ObjectDB = function(database, objectClass, debug) {
+
+    let ObjectDB = function(database, objectClass, debug) {
         this.db = database;
         this.setClass(objectClass);
         this.debugMode = debug || false;
     }
+
     ObjectDB.prototype.setClass = function(aClass) {
         this.objectClass = aClass;
         this.objectClassName = (new aClass()).getLabel();
-        this.byNamePath = '/' +  this.objectClassName + '/names';
-        this.byIdPath = '/' +  this.objectClassName + '/ids';
-        this.nextIdPath = '/' +  this.objectClassName + '/nextid';
-        this.uniqueIdPath = '/' +  this.objectClassName + '/unique';
-        this.debug('setClass(aClass='+aClass.name+'): this.byIdPath='+this.byIdPath);
+        this.byNamePath = '/' + this.objectClassName + '/names';
+        this.byIdPath = '/' + this.objectClassName + '/ids';
+        this.nextIdPath = '/' + this.objectClassName + '/nextid';
+        this.uniqueIdPath = '/' + this.objectClassName + '/unique';
+        this.debug('setClass(aClass=' + aClass.name + '): this.byIdPath=' + this.byIdPath);
     }
 
     ObjectDB.prototype.debug = function(msg) {
@@ -25,86 +26,79 @@
         this.debug('setUnique(' + attrNames + ') -> ' + this.uniqueIdPath + '/');
     }
 
+    /*
+     *
+     */
     ObjectDB.prototype.save = function(obj) {
-        var mustUpdateNextIdCounter = false;
-        var nextId = null;
         var uniqueConstraint = null;
         var reqStatus = 'success';
         var reqStatusReason = null;
         var prevName;
-        try {
-            // check if unique constraint is respected
-            //uniqueConstraint = this.db.getData(this.uniqueIdPath);
-            //console.log('uniqueConstraint: '+uniqueConstraint);
-            //console.log('searching name: ' + obj.name);
-            var dataId = this.getIdForName(obj.name);
-            // Check name not already exist
-            //console.log('getIdForName(' + obj.name + ')=' + dataId);
-            if (dataId != null) {
-                if (obj.id != dataId) {
-                    reqStatus = 'error'
-                    reqStatusReason = "data with name: " + obj.name + " already exists";
-                    this.debug('save(obj name=' + obj.name + ') NAME ALREADY EXISTS');
-                }
-            }
-        } catch (error) {
-            //console.error('### ERROR: getIdForName - ' + error.message);
-        }
 
-        if (reqStatusReason == null) { // No duplication attempt
-            //get new next id
+        // Check object ID / must be provided even for a creation
+        if (obj.id == null || obj.id == undefined) {
+            // No ID is provided
+            reqStatus = 'error'
+            reqStatusReason = "data without id : " + obj.name + " - id is mandatory even for object creation;";
+            this.debug('save(obj name=' + obj.name + ') NO ID PROVIDED');
+        } else {
+            // ID is provided
+            // Check if another object exist with the same name and a different ID
             try {
-                nextId = this.db.getData(this.nextIdPath);
-                if (obj.id == null) {
-                    // new Object
-                    obj.id = nextId;
-                    nextId++;
-                    mustUpdateNextIdCounter = true;
-                    this.debug('save(obj name=' + obj.name + ') NEW OBJECT id=' + obj.id);
-                } else if (obj.id >= nextId) {
-                    nextId = obj.id + 1;
-                    mustUpdateNextIdCounter = true;
-                    this.debug('save(obj name=' + obj.name + ') NEW OBJECT WITH id=' + obj.id + ' > MAXID');
+                var data = this.getForName(obj.name);
+            } catch (error) {
+                console.error('### ERROR: getIdForName - ' + error.message);
+            }
+            try {
+                if (data != null && data.id != undefined && obj.id != data.id) {
+                    // Another Object found with different ID
+                    reqStatus = 'error';
+                    reqStatusReason = 'data with name: ' + obj.name + ', ID: ' + data.id + ' already exists.';
+                    this.debug('save(obj name=' + obj.name + ', id=' + data.id + ') NAME ALREADY EXISTS');
+                } else {
+                    data = this.getForId(obj.id);
+                    if (data != null) {
+                        // Update case
+                        // Check for object version integrity : 
+                        //    given object version and database version must be equal
+                        if (data.version != obj.version) {
+                            // Version integrity issue
+                            reqStatus = 'error';
+                            reqStatusReason = 'version: ' + obj.version + ' for data with name: ' + obj.name + ', ID: ' + data.id + ' doesn\'t match database one :' + data.version;
+                            this.debug('save(obj name=' + obj.name + ', id=' + data.id + ') VERSION INTEGRITY ISSUE');
+                        } else {
+                            // Remove existing name index
+                            obj.version = obj.version + 1;
+                            this.db.delete(this.byNamePath + '/' + data.name);
+                            this.debug('save(obj name=' + obj.name + ') UPDATE (Version=' + obj.version + ')');
+                        }
+                    } else {
+                        // Create case
+                        this.debug('save(obj name=' + obj.name + ') CREATE');
+                    }
+                    this.db.push(this.byIdPath + '/' + obj.id, obj);
+                    this.db.push(this.byNamePath + '/' + obj.name, obj.id);
                 }
             } catch (error) {
-                mustUpdateNextIdCounter = true;
-                if (obj.id == null) {
-                    // not found - new object type => id=1
-                    obj.id = 1;
-                    this.debug('save(obj name=' + obj.name + ') INIT MAXID = 2');
-                }
-                nextId = obj.id + 1;
-            }
-
-            var data = this.getForId(obj.id);
-            //obj must have a name and id attribute
-            // store all data by id 
-            this.db.push(this.byIdPath + '/' + obj.id, obj);
-
-            // store or update an index with name/id
-            if (data != null) {
-                this.debug('save(obj name=' + obj.name + ') UPDATE');
-                this.db.delete(this.byNamePath + '/' + data.name);
-            } else {
-                this.debug('save(obj name=' + obj.name + ') CREATE');
-            }
-            this.db.push(this.byNamePath + '/' + obj.name, obj.id);
-
-            if (mustUpdateNextIdCounter) {
-                // store new next index value
-                this.db.push(this.nextIdPath, nextId);
-
-                this.debug('save() UPDATE MAXID = ' + nextId);
+                console.error('### ERROR: getIdForName - ' + error.message);
             }
         }
-        return { status: reqStatus, message: reqStatusReason, id: obj.id };
-    };
+        return {
+            status: reqStatus,
+            message: reqStatusReason,
+            id: obj.id,
+            version: obj.version
+        };
+    }
 
+
+    /*
+     *
+     */
     ObjectDB.prototype.getAll = function() {
         var data = null;
         try {
-
-            this.debug('this.db.getData('+this.byIdPath+')');
+            this.debug('this.db.getData(' + this.byIdPath + ')');
             data = this.db.getData(this.byIdPath);
         } catch (error) {
             this.debug('### ERROR: getAll - ' + error.message);
@@ -131,7 +125,7 @@
         try {
             data = this.db.getData(this.byIdPath + '/' + id);
         } catch (error) {
-            //console.error('### ERROR: getForId - ' + error.message);
+            console.debug('### ERROR: getForId - ' + error.message);
         }
         var obj = null;
         if (data != null) {
@@ -143,7 +137,8 @@
         this.debug('getForId(' + id + ')' + (data == null ? ' NOT FOUND' : ' OK'));
         return data;
     };
-    /**
+
+    /*
      * Get the Name stored in the data base for the given Application id
      * @param id of the searched application
      * @returns name
@@ -153,13 +148,15 @@
         try {
             data = this.db.getData(this.byNamePath + '/' + name);
         } catch (error) {
-            //console.error('### ERROR: getIdForName - ' + error.message);
+            console.debug('### ERROR: getIdForName - ' + error.message);
         }
-
         this.debug('getForIdName(' + name + ')' + (data == null ? ' NOT FOUND' : ' OK'));
         return data;
     };
 
+    /*
+     *
+     */
     ObjectDB.prototype.getForName = function(name) {
         var data = null;
         try {
@@ -169,7 +166,7 @@
             }
         } catch (error) {
             data = null;
-            //console.error('### ERROR: getForName - ' + error.message);
+            console.debug('### ERROR: getForName - ' + error.message);
         }
         var obj = null;
         if (data != null) {
@@ -181,20 +178,31 @@
         return data;
     };
 
-    ObjectDB.prototype.delete = function(objId) {
+    ObjectDB.prototype.delete = function(objId, objVersion) {
         var reqStatus = 'success';
         var reqStatusReason = null;
         try {
             var obj = this.getForId(objId);
-            this.db.delete(this.byNamePath + '/' + obj.name);
-            this.db.delete(this.byIdPath + '/' + obj.id);
+            if (obj.version == objVersion) {
+                this.db.delete(this.byNamePath + '/' + obj.name);
+                this.db.delete(this.byIdPath + '/' + obj.id);
+            } else {
+                reqStatus = 'error';
+                reqStatusReason = 'trying to delete object id=' + objId + ' with a version=' + objVersion + ' that does\'t match the database one=' + obj.version;
+                this.debug('### ERROR: delete - ' + reqStatusReason);
+            }
         } catch (error) {
-            //console.error('### ERROR: delete - ' + error.message);
+            console.error('### ERROR: delete - ' + error.message);
             reqStatus = 'error';
             reqStatusReason = 'object id=' + objId + ' not found';
         }
         this.debug('delete(' + objId + ')' + (reqStatusReason == null ? ' OK' : ' NOT FOUND'));
-        return { status: reqStatus, message: reqStatusReason, id: obj.id };
+        return {
+            status: reqStatus,
+            message: reqStatusReason,
+            id: obj.id,
+            version: obj.version
+        };
     }
 
     module.exports = ObjectDB;
